feat: expose live phase 1 demo endpoint
