+++ src/ootils_core/api/app.py
@@ -19,7 +19,7 @@ from starlette.middleware.base import BaseHTTPMiddleware
 
 from ootils_core.api.auth import _expected_token
 from ootils_core.api.dependencies import _get_ootils_db, get_db
-from ootils_core.api.routers import bom, calc, calendars, dq, events, explain, forecasting, ghosts, graph, ingest, issues, mrp, mrp_apics, planning_params, projection, rccp, scenarios, simulate
+from ootils_core.api.routers import bom, calc, calendars, demo, dq, events, explain, forecasting, ghosts, graph, ingest, issues, mrp, mrp_apics, planning_params, projection, rccp, scenarios, simulate
 from ootils_core.api.routers.graph import nodes_router
 from ootils_core.mps import router as mps_router
 from ootils_core.atp import atp_router
@@ -233,6 +233,7 @@ def create_app() -> FastAPI:
     application.include_router(planning_params.router)
     application.include_router(scenarios.router)
     application.include_router(calc.router)
+    application.include_router(demo.router)
     application.include_router(mrp.router)
     application.include_router(mrp_apics.router)
     application.include_router(forecasting.router)
