+++ src/ootils_core/demo/__init__.py
@@ -1,0 +1,1 @@
+"""Demo flows for Ootils Core."""
